Front End Requirements 1. Must be built with Angular 17.x 2. Create a Landing Page where the user can click 3 buttons: a. Start Approval b. Approve c. Reject Back End Requirements 1. Azure Function App that runs in C# 2. Create Http Triggers than can respond to the 3 actions noted on the front end: Technical Interview
PROPERTY OF CONTRACTS 365, INC. | CONFIDENTIAL INFORMATION | DO NOT DISTRIBUTE 2

a. Start Approval
i. An Orchestration is started that waits for an Approval signal. The
orchestration should not complete until another action is taken. It could sit
there for days or months if needed. An email is sent to the initiated user that
the approval started.

b. Approve
i. If approved button is clicked, then orchestration would complete, and an
email would be sent to the initiated user that the approval has been
approved.
c. Rejected
i. If the rejected button is clicked, then an email goes out to the initiated user to
state that the approval was rejected.